requirements.txt
fastapi
uvicorn
scikit-learn
pandas
python-dotenv
joblib
matplotlib
seaborn 